pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "astrapy>=2.0.1",
    "llama-index>=0.12.42",
    "numpy>=2.3.0",
    "openai>=1.86.0",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "pymupdf>=1.26.1",
    "pymupdf4llm>=0.0.25",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "streamlit>=1.45.1",
    "tqdm>=4.67.1",
    "upstash-vector>=0.8.0",
    "voyageai>=0.3.2",
]
